Fixed MinRadius Issue (#211)
* Fixed MinRadius Issue

No longer sends players to exclusively the corners of the map. Now sends them to anywhere where x or z is above the MinRadius (while both being below the MaxRadius)

* Fixed MinRadius Issue

No longer sends players to exclusively the corners of the map. Now sends them to anywhere where x or z is above the MinRadius (while both being below the MaxRadius)


Fixes this issue.
https://github.com/RonanPlugins/BetterRTP/issues/210

* Update QueueHandler.java

* MaxRadius and MinRadius fix